create 3 components
